Move existing serialization example to `json` sub-folder
Motivation:

Prepare `servicetalk-examples/http/serialization` for more examples for
different formats.

Modification:

- Move `servicetalk-examples:http:serialization` to
`servicetalk-examples:http:serialization:json`;
- Rename module to `servicetalk-examples-http-serialization-json`;
- Update references in docs;

Result:

`servicetalk-examples/http/serialization` can be used to add more
examples.
