--- servicetalk-examples/http/serialization/build.gradle
@@ -1,26 +1,0 @@
-/*
- * Copyright © 2019 Apple Inc. and the ServiceTalk project authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-apply plugin: "java"
-apply from: "../../gradle/idea.gradle"
-
-dependencies {
-  implementation project(":servicetalk-annotations")
-  implementation project(":servicetalk-data-jackson")
-  implementation project(":servicetalk-http-netty")
-
-  runtimeOnly "org.apache.logging.log4j:log4j-slf4j-impl:$log4jVersion"
-}
