+++ servicetalk-examples/http/serialization/json/src/main/resources/log4j2.xml
